refactor: security checks

+++ pkg/cmd/create/domain/domain.go
@@ -22,7 +22,7 @@ type Fields struct {
 	Name                 string   `json:"name"`
 	Cnames               []string `json:"cnames"`
 	CnameAccessOnly      string   `json:"cname_access_only"`
-	EdgeApplicationID    int      `json:"edge_application_id"`
+	EdgeApplicationID    int64    `json:"edge_application_id"`
 	DigitalCertificateID string   `json:"digital_certificate_id"`
 	IsActive             string   `json:"is_active"`
 	Path                 string
@@ -63,7 +63,7 @@ func NewCmd(f *cmdutil.Factory) *cobra.Command {
 						return msg.ErrorConvertApplicationID
 					}
 
-					fields.EdgeApplicationID = int(num)
+					fields.EdgeApplicationID = num
 				}
 
 				if !cmd.Flags().Changed("name") {
@@ -89,7 +89,7 @@ func NewCmd(f *cmdutil.Factory) *cobra.Command {
 
 				request.SetName(fields.Name)
 				request.SetCnames(fields.Cnames)
-				request.SetEdgeApplicationId(int64(fields.EdgeApplicationID))
+				request.SetEdgeApplicationId(fields.EdgeApplicationID)
 
 				if cmd.Flags().Changed("digital-certificate-id") {
 					request.SetDigitalCertificateId(fields.DigitalCertificateID)
@@ -122,7 +122,7 @@ func NewCmd(f *cmdutil.Factory) *cobra.Command {
 	flags.StringSliceVar(&fields.Cnames, "cnames", []string{}, msg.FlagCnames)
 	flags.StringVar(&fields.CnameAccessOnly, "cname-access-only", "false", msg.FlagCnameAccessOnly)
 	flags.StringVar(&fields.DigitalCertificateID, "digital-certificate-id", "", msg.FlagDigitalCertificateID)
-	flags.IntVar(&fields.EdgeApplicationID, "application-id", 0, msg.FlagEdgeApplicationId)
+	flags.Int64Var(&fields.EdgeApplicationID, "application-id", 0, msg.FlagEdgeApplicationId)
 	flags.StringVar(&fields.IsActive, "active", "true", msg.FlagIsActive)
 	flags.StringVar(&fields.Path, "file", "", msg.FlagFile)
 	flags.BoolP("help", "h", false, msg.HelpFlag)
